Fixed npm to install bootstrap, rename action types, VehicleOne to VehicleDetails, app/views/partials to app/components, views to layouts
