+++ package.json
@@ -17,13 +17,16 @@
     "cross-env": "^5.0.1",
     "extract-text-webpack-plugin": "^3.0.0",
     "laravel-mix": "^1.0",
-    "lodash": "^4.17.4"
+    "lodash": "^4.17.4",
+    "url-loader": "^0.5.9"
   },
   "dependencies": {
     "babel-polyfill": "^6.23.0",
+    "bootstrap": "^4.0.0-alpha.6",
     "prop-types": "^15.5.10",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
+    "react-fa": "^4.2.0",
     "react-redux": "^5.0.5",
     "react-router": "^3.0.0",
     "redux": "^3.7.2",
