+++ resources/assets/js/app/bootstrap/bootstrap.scss
@@ -1,0 +1,2 @@
+@import "./variables";
+@import "~bootstrap/scss/bootstrap";
