+++ resources/assets/js/app/components/MainHeader.js
