+++ resources/assets/js/app/components/PageHeader.js
