+++ resources/assets/js/app/layouts/Home.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import PageHeader from './partials/PageHeader';
+import PageHeader from '../components/PageHeader';
 
 class Home extends Component {
