+++ resources/assets/js/features/vehicle/actionTypes.js
@@ -1,8 +1,6 @@
-export const GET_VEHICLES = 'GET_VEHICLES';
-export const GET_VEHICLE = 'GET_VEHICLE';
+export const VEHICLE_GET_ALL = 'VEHICLE_GET_ALL';
+export const VEHICLE_GET = 'VEHICLE_GET';
 
-export const ADD_VEHICLE = 'ADD_VEHICLE';
-export const EDIT_VEHICLE = 'EDIT_VEHICLE';
-export const DELETE_VEHICLE = 'DELETE_VEHICLE';
-
-export const CHANGE_FILTER_VALUE = 'CHANGE_FILTER_VALUE';
+export const VEHICLE_ADD = 'VEHICLE_ADD';
+export const VEHICLE_EDIT = 'VEHICLE_EDIT';
+export const VEHICLE_DELETE = 'VEHICLE_DELETE';
