+++ resources/assets/js/features/vehicle/layouts/Vehicles.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import PageHeader from '../../../app/views/partials/PageHeader';
+import PageHeader from '../../../app/components/PageHeader';
 import VehiclesList from '../components/VehiclesList';
 
 import '../styles/vehicle.scss';
